fix(MatchResult): matchCount 타입 long으로 수정
- Lotto 클래스의 countMatchNumbers메서드 반환 타입에 맞춤
- count() 반환 타입이 long

+++ src/main/java/lotto/domain/MatchResult.java
@@ -1,15 +1,15 @@
 package lotto.domain;
 
 public class MatchResult {
-    private final int matchCount;
+    private final long matchCount;
     private final boolean matchBonus;
 
-    public MatchResult(int matchCount, boolean matchBonus) {
+    public MatchResult(long matchCount, boolean matchBonus) {
         this.matchCount = matchCount;
         this.matchBonus = matchBonus;
     }
 
-    public int getMatchCount() {
+    public long getMatchCount() {
         return matchCount;
     }
